Fix critical Preset E encoding bug - loops now work!
Root cause: TARREG and SETBYTE were losing bits 9-8 during decode, causing:
1. Target register to be extracted from wrong bits
2. Immediate values in SETBYTE to include target register bits
3. Loop counters to get stuck at value 1

## Changes

### PPU Fixes
- Changed executePresetE() to accept uint16_t (10 bits) instead of uint8_t (8 bits)
- Fixed Preset E extraction to pass bits 9-0 instead of bits 7-0
- Updated TARREG decoding:
  - Target register: bits 9-8 (was incorrectly 7-6)
  - Byte select: bit 7 (was incorrectly bit 5)
  - Source register: bits 5-0 (unchanged)
- Updated SETBYTE decoding:
  - Target register: bits 9-8 (was incorrectly 7-6)
  - Immediate value: bits 7-0 (was incorrectly full 8 bits including target)

Files modified:
- include/ppu.h: executePresetE signature changed to uint16_t
- src/ppu.cpp: Fixed Preset E extraction and TARREG/SETBYTE decoding
- tools/test_demo.cpp: Disabled HLT detection temporarily for testing

## Test Results

Simple loop test (increment R10 from 0 to 5):
- ✅ R10 correctly increments: 0 → 1 → 2 → 3 → 4 → 5 → 6
- ✅ Loop exits when R10 > R15 (6 > 5)
- ✅ PC register (R62) correctly set to loop address (14)

The fundamental loop bug is FIXED! 🎉

Next steps: Update assembler to match corrected encoding.

+++ include/ppu.h
@@ -174,7 +174,7 @@ private:
 
     // Execution methods
     void executeInstruction();
-    void executePresetE(uint8_t subopcode, uint8_t operand);
+    void executePresetE(uint8_t subopcode, uint16_t operand);
     void executePresetF(uint8_t subopcode, uint8_t operand);
 
     // Fetch instruction
